remove lstrip_blocks option from render_template
This was apparently introduced in jinja 2.7 which is far
beyond our minimum supported targets for holland2. i.e.
rhel6 has jinja 2.2.1 and rhel5+epel has jinja 2.5.5.

This is worked around by adjusting the template code, which
we'll do for now.  This is only a pretty-print option and
does not currently affect functionality.