Fixed socket checking code in file backup
Apparently, python's way of doing an "os.path.join" confused me in it's
execution and I thought that the code worked without fault. I was wrong.
If you specify another root path after the first path, it will clear the
current path and replace it with the current one, preventing me from
seeing my error in the first place.